// Ref/Web/ReferenceDataUpdateService.Web/src/models/RefStlScriptUserView.ts
//------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated from a template.
//
//     Manual changes to this file may cause unexpected behavior in your application.
//     Manual changes to this file will be overwritten if the code is regenerated.
// </auto-generated>
//------------------------------------------------------------------------------
import { IEntity } from "./IEntity";

export class RefStlScriptUserView implements IEntity {
	STL_PK: string;
	STL_FeatureCode: string;
	static readonly STL_FeatureCode_MaxLength: number = 3;
	STL_RoleName: string;
	static readonly STL_RoleName_MaxLength: number = 50;
	STL_ModuleName: string;
	static readonly STL_ModuleName_MaxLength: number = 50;
	STL_FunctionName: string;
	static readonly STL_FunctionName_MaxLength: number = 50;
	STL_FeatureName: string;
	static readonly STL_FeatureName_MaxLength: number = 75;
	STL_DataGranularity: string;
	static readonly STL_DataGranularity_MaxLength: number = 3;
	STL_CompanyCode: string;
	static readonly STL_CompanyCode_MaxLength: number = 1000;
	STL_BranchCode: string;
	static readonly STL_BranchCode_MaxLength: number = 1000;
	STL_TransactionDateUtc: string;
	static readonly STL_TransactionDateUtc_MaxLength: number = 1000;
	STL_CreatingUserCode: string;
	static readonly STL_CreatingUserCode_MaxLength: number = 1000;
	STL_GuidReference: string;
	static readonly STL_GuidReference_MaxLength: number = 1000;
	STL_BillingReference1: string;
	static readonly STL_BillingReference1_MaxLength: number = 1000;
	STL_BillingReference2: string;
	static readonly STL_BillingReference2_MaxLength: number = 1000;
	STL_BillingReference3: string;
	static readonly STL_BillingReference3_MaxLength: number = 1000;
	STL_BillingReference4: string;
	static readonly STL_BillingReference4_MaxLength: number = 1000;
	STL_AdditionalRefs: string;
	static readonly STL_AdditionalRefs_MaxLength: number = 1000;
	STL_TransactionCount: string;
	static readonly STL_TransactionCount_MaxLength: number = 1000;
	STL_PreparationScript: string;
	STL_FromClause: string;
	STL_WhereClause: string;
	STL_WithOptionRecompile: boolean;
	STL_UsedInBilling: boolean;
	STL_ActiveOn: string;
	static readonly STL_ActiveOn_MaxLength: number = 3;
	STL_MinCW1Version: string;
	static readonly STL_MinCW1Version_MaxLength: number = 20;
	STL_MaxCW1Version: string;
	static readonly STL_MaxCW1Version_MaxLength: number = 20;
	STL_DateType: string;
	static readonly STL_DateType_MaxLength: number = 3;
	STL_CollectionStartDateUtc: string | null;
	STL_IsSystem: boolean;
	STL_IsPublished: boolean;
	STL_IsEditable: boolean;
}